calculationserver/requirements.txt
flask==0.12.2
gunicorn==19.7.1
nltk
python-rake
